Switched to pytest
